dcp-t5(003): move cli/worker_sessions under worker_sessions
Relocate pkg/transports/cli/worker_sessions to
pkg/services/worker_sessions/transports/cli/worker_sessions. The package
imports only the worker_sessions CLI service among product services, so
it belongs in that service's own transports subtree.

Pure relocation: no file contents change beyond import paths in six
external importers (cli root composition and pkg/wire, including the
generated wire_gen.go whose import line is retargeted in step with
wire.go so the pair stays drift-free). Existing import aliases already
described the package rather than its location, so they are unchanged.
Unit and functional coverage minimums move with the package.
