+++ docs/internal/baselines/go-functional-coverage-package-minimums.json
@@ -1168,6 +1168,10 @@
       "minimum": 39.01
     },
     {
+      "package": "github.com/portpowered/infinite-you/pkg/services/worker_sessions/transports/cli/worker_sessions",
+      "minimum": 0.00
+    },
+    {
       "package": "github.com/portpowered/infinite-you/pkg/services/worker_sessions/transports/http",
       "minimum": 52.16
     },
@@ -1352,10 +1356,6 @@
     {
       "package": "github.com/portpowered/infinite-you/pkg/transports/cli/work",
       "minimum": 27.70
-    },
-    {
-      "package": "github.com/portpowered/infinite-you/pkg/transports/cli/worker_sessions",
-      "minimum": 0.00
     },
     {
       "package": "github.com/portpowered/infinite-you/pkg/transports/http",
