+++ pkg/services/worker_sessions/transports/cli/worker_sessions/list.go
